+++ test/controllers/home_controller_test.rb
@@ -14,7 +14,7 @@ class HomeControllerTest < ActionController::TestCase
     @current_subscription = nil
   end
 
-  test 'non-logged in user is redirected to the login page' do
+  test 'non-logged in user is redirected to the login page if no hmac and shop domain present' do
     log_out
     get(:index)
     assert_redirected_to ShopifyApp::Engine.routes.url_helpers.login_path
@@ -87,6 +87,15 @@ class HomeControllerTest < ActionController::TestCase
     @shop.installed!
     get(:index)
     assert_response :success
+  end
+
+  test 'non-logged in user is logged in if valid hmac and shop domain is present in url' do
+    log_out
+    Timecop.freeze('2017-03-08 12:44:58 +1100') do
+      hmac = 'eb49ba93a8daf8a11a04c66129faf98de1cd40f082b0ae78e79a2dfbbefb438d'
+      get(:index, { hmac: hmac, shop: 'widgets-dev.myshopify.com', timestamp: Time.now.to_i })
+      assert_response  :success
+    end
   end
 
 end